refactor to support multiple os
