+++ .gitignore
@@ -1,5 +1,6 @@
 # KubeArmor
 kubearmor
+KubeArmor.exe
 karmor
 snitch
 kubearmor-operator
