+++ KubeArmor/common/boot_time_win.go
@@ -1,0 +1,8 @@
+//go:build windows
+
+package common
+
+func GetBootTime() string {
+	// TODO: get boot time in windows
+	return ""
+}
