+++ KubeArmor/common/common.go
@@ -24,7 +24,6 @@ import (
 	kc "github.com/kubearmor/KubeArmor/KubeArmor/config"
 	kg "github.com/kubearmor/KubeArmor/KubeArmor/log"
 	tp "github.com/kubearmor/KubeArmor/KubeArmor/types"
-	"golang.org/x/sys/unix"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -621,19 +620,6 @@ func HandleGRPCErrors(err error) error {
 	}
 
 	return nil
-}
-
-// get boot time
-// credits: https://github.com/kubernetes/kubernetes/blob/master/pkg/kubelet/util/boottime_util_linux.go
-func GetBootTime() string {
-	currentTime := time.Now()
-
-	var info unix.Sysinfo_t
-	if err := unix.Sysinfo(&info); err != nil {
-		return ""
-	}
-
-	return currentTime.Add(-time.Duration(info.Uptime) * time.Second).Truncate(time.Second).UTC().String()
 }
 
 func GetLabelsFromString(labelString string) (map[string]string, []string) {
