+++ KubeArmor/enforcer/bpflsm/enforcer.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 // SPDX-License-Identifier: Apache-2.0
 // Copyright 2022 Authors of KubeArmor
