+++ KubeArmor/feeder/helpers.go
@@ -1,0 +1,21 @@
+//go:build linux
+
+package feeder
+
+import (
+	"os"
+	"strconv"
+	"syscall"
+)
+
+func getFileProcessUID(path string) string {
+	info, err := os.Stat(path)
+	if err == nil {
+		stat := info.Sys().(*syscall.Stat_t)
+		uid := stat.Uid
+
+		return strconv.Itoa(int(uid))
+	}
+
+	return ""
+}
