+++ KubeArmor/feeder/policyMatcher.go
@@ -5,12 +5,10 @@ package feeder
 
 import (
 	"math"
-	"os"
 	"path/filepath"
 	"regexp"
 	"strconv"
 	"strings"
-	"syscall"
 
 	cfg "github.com/kubearmor/KubeArmor/KubeArmor/config"
 	tp "github.com/kubearmor/KubeArmor/KubeArmor/types"
@@ -86,18 +84,6 @@ func fetchProtocol(resource string) string {
 		return "packet"
 	}
 	return resource
-}
-
-func getFileProcessUID(path string) string {
-	info, err := os.Stat(path)
-	if err == nil {
-		stat := info.Sys().(*syscall.Stat_t)
-		uid := stat.Uid
-
-		return strconv.Itoa(int(uid))
-	}
-
-	return ""
 }
 
 // getOperationAndCapabilityFromName Function
