+++ KubeArmor/main.go
@@ -7,6 +7,7 @@ package main
 import (
 	"os"
 	"path/filepath"
+	"runtime"
 
 	"github.com/kubearmor/KubeArmor/KubeArmor/buildinfo"
 	cfg "github.com/kubearmor/KubeArmor/KubeArmor/config"
@@ -19,7 +20,7 @@ func init() {
 }
 
 func main() {
-	if os.Geteuid() != 0 {
+	if runtime.GOOS != "windows" && os.Geteuid() != 0 {
 		if os.Getenv("KUBEARMOR_UBI") == "" {
 			kg.Printf("Need to have root privileges to run %s\n", os.Args[0])
 			return
