+++ KubeArmor/presets/protectproc/preset.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 // SPDX-License-Identifier: Apache-2.0
 // Copyright 2021 Authors of KubeArmor
 
@@ -17,6 +19,7 @@ import (
 	"github.com/cilium/ebpf/ringbuf"
 	"github.com/cilium/ebpf/rlimit"
 	"github.com/kubearmor/KubeArmor/KubeArmor/buildinfo"
+	"github.com/kubearmor/KubeArmor/KubeArmor/common"
 	"github.com/kubearmor/KubeArmor/KubeArmor/presets/base"
 
 	fd "github.com/kubearmor/KubeArmor/KubeArmor/feeder"
@@ -106,12 +109,12 @@ func (p *Preset) RegisterPreset(logger *fd.Feeder, monitor *mon.SystemMonitor) (
 		Pinning:    ebpf.PinByName,
 		Name:       "protectproc_preset_containers",
 	}, ebpf.MapOptions{
-		PinPath: monitor.PinPath,
+		PinPath: common.GetMapRoot(),
 	})
 
 	if err := loadProtectprocObjects(&p.obj, &ebpf.CollectionOptions{
 		Maps: ebpf.MapOptions{
-			PinPath: monitor.PinPath,
+			PinPath: common.GetMapRoot(),
 		},
 	}); err != nil {
 		p.Logger.Errf("error loading BPF LSM objects: %v", err)
