+++ KubeArmor/utils/bpflsmprobe/probe.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 // SPDX-License-Identifier: Apache-2.0
 // Copyright 2023 Authors of KubeArmor
